Move scoring related tests to new test file
